requirements.txt
# Core Dependencies
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality (optional)
flake8==6.1.0
black==23.12.0
mypy==1.7.1

# Development (optional)
ipython==8.18.1
ipdb==0.13.13